started genelib and moved suffix checker there
